Setup a fall back for when a video reports a nonsensical fps value (like 16,000 in one case.)
Set a default value for duration when a video does not report it's duration.
This value is used informationally, but not logically so we can just make
up anything for now.

Create a compromise (between fit/zoom) for putting a portrait video in
a landscape layout or visa versa.

When writing out trimmed/padded aligned video clips, downsample them a
bit just to keep cumulative sizes from growing insane.